Refactor SchameUiHandler to use WidgetRouter.
