+++ modules/json_form_widget/json_form_widget.services.yml
@@ -15,6 +15,9 @@ services:
   json_form.array_helper:
     class: \Drupal\json_form_widget\ArrayHelper
     arguments: ['@json_form.object_helper']
+  json_form.widget_router:
+    class: \Drupal\json_form_widget\WidgetRouter
+    arguments: ['@uuid','@json_form.string_helper', '@dkan.metastore.service']
   json_form.schema_ui_handler:
     class: \Drupal\json_form_widget\SchemaUiHandler
-    arguments: ['@dkan.metastore.schema_retriever', '@logger.factory', '@uuid', '@json_form.string_helper', '@dkan.metastore.service']
+    arguments: ['@dkan.metastore.schema_retriever', '@logger.factory', '@json_form.widget_router']
